feat: 2d camera and many other changes

+++ engine/src/main/kotlin/gfx/TextureHandle.kt
@@ -2,4 +2,4 @@ package gfx
 
 // value classes are really cool! typesafe typealias
 @JvmInline
-value class TextureHandle(val id: Short)
+value class TextureHandle(val id: Int)
